backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-data-platform-backend"
version = "1.0.0"
description = "AI Data Analysis Platform Backend Services"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "AI Data Platform Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # FastAPI and modern web framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Flask (legacy support)
    "flask>=3.0.0",
    "Flask-CORS>=4.0.0",
    "Flask-JWT-Extended>=4.5.3",
    "Flask-SQLAlchemy>=3.0.5",
    
    # Database
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.7",
    "alembic>=1.12.0",
    
    # Authentication and Security
    "PyJWT>=2.8.0",
    "bcrypt>=4.0.1",
    "cryptography>=41.0.4",
    "python-multipart>=0.0.6",
    
    # HTTP and networking
    "httpx>=0.25.2",
    "requests>=2.31.0",
    
    # Data processing and analysis
    "pandas>=2.0.3",
    "numpy>=1.24.4",
    "PyMySQL>=1.1.0",
    
    # Environment and configuration
    "python-dotenv>=1.0.0",
    
    # Logging and monitoring
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]
mongo = [
    "pymongo>=4.5.0",
]
redis = [
    "redis>=4.6.0",
]
excel = [
    "openpyxl>=3.1.2",
]
mysql = [
    "PyMySQL>=1.1.0",
]
compute = [
    "ray[default]>=2.8.0",
    "dask[complete]>=2023.10.1",
    "celery[redis]>=5.3.4",
]

[project.urls]
Homepage = "https://github.com/your-org/ai-data-platform"
Repository = "https://github.com/your-org/ai-data-platform.git"
Issues = "https://github.com/your-org/ai-data-platform/issues"

[tool.setuptools]
packages = ["apps", "shared"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"